requirements.txt
streamlit==1.28.0
neo4j==5.15.0
zhipuai==2.0.1
bcrypt==4.0.1
Pillow==10.0.0
dashscope==1.14.0
